Fix pet rename triggering twice through menu
The `name` menu item would open the pet selection menu even if a pet was already selected, erroneously resulting in the rename command being run twice, and capturing the player's chat input twice. This was due to a missing early return in the item's onClick handler. Adding that return, in line with how it's done in the other menu items, fixes this!